refactor: renaming and minor restructuring
