+++ builder.php
@@ -5,7 +5,7 @@ require_once __DIR__ . "/vendor/autoload.php";
 use Src\Utils\CommonUtils;
 use Src\Utils\BuilderUtils;
 
-pcntl_alarm(10 * 60); // Kills script if it's taking too long.
+pcntl_alarm(10 * 60); // Kills job if it's taking too long.
 
 // user param
 $forceRebuild = in_array("-y", $argv);
