+++ src/Jobs/servidores.php
@@ -2,11 +2,11 @@
 
 require_once __DIR__ . "/../../vendor/autoload.php";
 
-use Src\Utils\ScriptUtils;
+use Src\Jobs\Runner\Runner;
 
-pcntl_alarm(25 * 60); // Kills script if it's taking too long.
+pcntl_alarm(25 * 60); // Kills job if it's taking too long.
 
 $tempTables = ['create_vinculosServidores_temp'];
 $tableGroups = ['ServidoresTables'];
 
-ScriptUtils::runScript($tempTables, $tableGroups);
+Runner::runJob($tempTables, $tableGroups);
